+++ R/fpp2.R
@@ -1,0 +1,7 @@
+# This package is based on the tidyverse package, hence the copyright
+# attribution to RStudio.
+
+#' @keywords internal
+"_PACKAGE"
+
+NULL